add purchase policy in the gui2

+++ trading-system-server/src/main/java/com/wsep202/TradingSystem/dto/PurchasePolicyDto.java
@@ -23,14 +23,12 @@ import java.util.Set;
 @Builder
 public class PurchasePolicyDto {
 
-    protected int purchaseId;
+    protected long purchaseId;
 
     /**
      * one of the following : 'all store', 'specific product', 'on system', 'on user',  'compose'
      */
     private String purchaseType;
-
-
 
     private String description;
     /**
